rafactored search api; completed yake respone to UI; added tests; refactored models.

+++ src/main/java/com/example/ArticleAI/modules/actualityResolver/models/Actuality.java
@@ -9,5 +9,5 @@ import lombok.*;
 @Setter
 public class Actuality {
     private int classId;
-    private Integer actuality;
+    private Long actuality;
 }
